[JENKINS-62297] Rename to UserCount
