+++ src/test/java/com/cloudbees/jenkins/support/impl/UserCountTest.java
@@ -1,11 +1,10 @@
 /*
  * Copyright © 2013 CloudBees, Inc.
  */
-package com.cloudbees.jenkins.support.timer;
+package com.cloudbees.jenkins.support.impl;
 
 import com.cloudbees.jenkins.support.SupportTestUtils;
 import com.cloudbees.jenkins.support.api.Component;
-import com.cloudbees.jenkins.support.impl.UsersContent;
 import hudson.ExtensionList;
 import hudson.model.User;
 import jenkins.security.LastGrantedAuthoritiesProperty;
@@ -17,7 +16,7 @@ import org.jvnet.hudson.test.JenkinsRule;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.MatcherAssert.assertThat;
 
-public class UsersContentTest {
+public class UserCountTest {
 
     @Rule
     public JenkinsRule j = new JenkinsRule();
@@ -31,8 +30,8 @@ public class UsersContentTest {
         User.getOrCreateByIdOrFullName("dave").addProperty(new LastGrantedAuthoritiesProperty());
         User.getOrCreateByIdOrFullName("eve").addProperty(new LastGrantedAuthoritiesProperty());
                 
-        String usersMdToString = SupportTestUtils.invokeComponentToString(ExtensionList.lookup(Component.class).get(UsersContent.class));
-        assertThat(usersMdToString, containsString(" * Non Authenticated Users count: " + 3));
-        assertThat(usersMdToString, containsString(" * Authenticated Users count: " + 2));
+        String usersMdToString = SupportTestUtils.invokeComponentToString(ExtensionList.lookup(Component.class).get(UserCount.class));
+        assertThat(usersMdToString, containsString(" * Non Authenticated User Count: " + 3));
+        assertThat(usersMdToString, containsString(" * Authenticated User Count: " + 2));
     }
 }
